requirements.txt
beautifulsoup4>=4.12.0
selenium>=4.15.0
python-dotenv>=1.0.0
pytest>=7.4.0
requests>=2.31.0 